added role status, in testing

+++ ophthalmology-backend/src/main/java/com/opth/clinic/ophthalmology_backend/service/DoctorService.java
@@ -3,6 +3,7 @@ package com.opth.clinic.ophthalmology_backend.service;
 import com.opth.clinic.ophthalmology_backend.exception.NotFoundException;
 import com.opth.clinic.ophthalmology_backend.model.Doctor;
 import com.opth.clinic.ophthalmology_backend.respository.DoctorRespository;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
@@ -18,11 +19,22 @@ public class DoctorService {
         return doctorRespository.findById(id).get();
     }
 
-    public Doctor addDoctor(Doctor doctor) {
+    public Doctor addDoctor(Doctor doctor, Authentication auth) {
+        String role = auth.getAuthorities().iterator().next().getAuthority();
+
+        if (!role.equals("ROLE_ADMIN")) {
+            throw new RuntimeException("Nie masz pozwolenia!");
+        }
         return doctorRespository.save(doctor);
     }
 
-    public Doctor updateDoctor(Long id, Doctor doctor) {
+    public Doctor updateDoctor(Long id, Doctor doctor, Authentication auth) {
+        String role = auth.getAuthorities().iterator().next().getAuthority();
+
+        if (!role.equals("ROLE_ADMIN")) {
+            throw new RuntimeException("Nie masz pozwolenia!");
+        }
+
         Doctor existingDoctor = doctorRespository.findById(id).orElseThrow(() -> new NotFoundException("Nie znaleziono lekarza"));
         if(doctor.getFirstName() != null) existingDoctor.setFirstName(doctor.getFirstName());
         if(doctor.getLastName() != null) existingDoctor.setLastName(doctor.getLastName());
@@ -30,7 +42,12 @@ public class DoctorService {
         return doctorRespository.save(existingDoctor);
     }
 
-    public void deleteDoctorById(Long id) {
+    public void deleteDoctorById(Long id, Authentication auth) {
+        String role = auth.getAuthorities().iterator().next().getAuthority();
+
+        if (!role.equals("ROLE_ADMIN")) {
+            throw new RuntimeException("Nie masz pozwolenia!");
+        }
         doctorRespository.deleteById(id);
     }
